contact/serializers.py
from rest_framework import serializers
from .models import ContactForm

class ContactFormSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactForm
        fields = ['first_name', 'last_name', 'company', 'email', 'message', 'created_at']