requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv 
gunicorn
requests
psycopg2-binary
whitenoise